Refactored gsheets.py and tested to confirm that Google Sheet is created. property_gsheets.py not yet functional. (Had to add os/sys imports to access property_level_analysis.py after siloing gsheets code in new folder.)
